Language changes related to Tetra slugs
The `AA_RechargingBatteries` hediff will now say "creature" instead of "TetraSlug".

The `AA_AchievementGreenestEnergy` achievement will now say "Have a friendly creature (like a tame Tetra Slug) recharge your batteries" (instead of "Tame a TetraSlug...").

The word "creature", as well changing the "Tame" part into "Have a friendly creature" were chosen as the hediff can be applied to non-animal Pawns like humanoids (for example due to Alpha Genes) or potentially player-controlled mechanoids (not sure if there's a mod like that at the moment, but it would be possible).

Additionally, the appearances of the words "Tetraslug"/"TetraSlug" were changed into "Tetra slug" for consistency with the majority of other text in the mod, as using 2 words instead of 1 is how this creature seems to be referred to in-game (specifically based on the label of Tetra slug's `ThingDef`, `PawnKindDef`, and `lifeStages`) - if not desired I can simply remove those changes.

Since there was no changes to code itself or any changes to have gameplay effect (only changes to descriptions), the changes were applied to both 1.4 and 1.3 - again, if not desired I can simply remove the changes to 1.3 from this PR.