Update animal XML
First, any place using `<ToxicEnvironmentResistance>` had its Biotech requirement removed, since it's a vanilla feature now.

Second, I changed `<defaultFactionType>` to `<defaultFactionDef>`. It was renamed in-game, so I thought it may be for the best to do it. The old name was kept as an alias in-game, which is why it still works.

Third, any black hive insectoids were changed so their `<PawnKindDef>` inherits from `InsectKindBase`, rather than `AnimalKindBase`. This means all the black hive insectoids will benefit from extra movement speed on insectoid terrain.

From animals using `BaseInsect2` or `AA_AlphaBaseInsect` as a base, I did NOT include changes to:
- Any of the 'lisks
- Barb slinger
- Chemfuel myrmidon
- Any other insects that aren't from black hive

I wasn't sure if you wanted them to benefit from insectoid terrain or not, so I just left them as-is.

Additional notes due to using `InsectKindBase`:
- It forces `<canArriveManhunter>` to false, so if any animal had it to true/unspecified, I forced it back to true
- It forces `<defaultFactionDef>` to `Insect`, so again - I forced it to null if it was unspecified before

If you don't want that, you may want to go and change it.